refactor(capture): extract pipeline + runCapture with DI fetcher
Behavior-preserving refactor. No API changes.

The 7-stage capture pipeline moves out of captureTool.ts:
- runCapture.ts: orchestrator (stage 1 fetch + 2-7 wiring + finally cleanup)
- pipeline.ts: pure stages 2-7 (style collapse → CSS rewrite → asset collect
  → UID assign → token rewrite → nextActions), no I/O, no DB

captureTool.ts shrinks to a ~50-line adapter: input validation, capability
gates, targetForScope translation, and a delegation to runCapture().

validateSelector moves to pipeline.ts since it has no I/O; runCapture.ts
re-exports it for backwards compat with captureTool.ts.

The safeFetcher is now only constructed when mode actually wants assets —
restricted modes (dom-only, styles-only) skip the network entirely. Same
for the getDraftSite token fetch. This is the regression guard for the
Task 2 mode-gate tests.

runCapture.ts accepts an injected CaptureFetcher so tests can stub
Playwright without launching a browser.

Existing tests pass (32 captureTool + 85 core/adapters = 117 total).
Lint clean.
